Use i32 everywhere for counters
We actually should use u32 internally, but:
because Kotlin doesn't (yet) have unsigned integers, it's always going to pass us signed values.
We also do all the error checking on the Rust side, so we need to receive the signed value to check for negativity and bail out.

We still don't do overflow checks though.