Implement start and end times for pings
The start time can be stored and persisted as a datetime metric.
The end time is only ever needed in-memory for the moment the ping is
assembled and stored.

We also store the object creation time in the Glean object itself (which
is used as the first start time), because:

* that was the easy way
* glean-ac does something similar (storing it in the PingMaker object)
* we don't have global metrics in the core library itself on which we
  could just set the date