1557339: Simplify Glean.isInitialized
On the Glean side, creation and initialization are handled in a single step
(in Glean::new).  Therefore, it is impossible to pass a handle to this object
to the Kotlin side before it has been initialized.  Therefore, on the Kotlin
side, we can assume that if we have a handle, we have an initialized Rust
object.