Implement a timespan metric
We rely on start and stop passed as an integer due to FFI requirement.

In a world where we wouldn't need a dispatch queue on the Kotlin side we
could rely on `Instant` and `Duration` only on the Rust side, making
this all much nicer.
Until then we take integer timestamps.