Use a closure to get the ping's name
When we switched to either a string name or a ping type object we
introduced a generic function, but didn't adjust the log message.
This results in the object's string representation to be used (which is
most likely just `PingType0x12345` or something along those lines).

The non-generic wrapper functions do know how to get to the name, so we
can use that knowledge and pass it in.