Update regex dependency and reduce its feature set.
The only regex in use is a pure ASCII-only match on a length-limited
string.
We don't need any extended feature of the regex library nor any
advanced performance improvements.
Depending on the new version allows us to reduce the feature set and
thus reduce the overall code size. This makes compilation of this crate
faster and the generated libraries smaller on all platforms.

This also removes the regex feature from the env_logger, as all we need
is some default logging during development/debugging/testing.