Introduce the GleanACDataMigrator.kt class
This class holds all the functions to perform
data migration from glean-ac to glean-core, on
the Kotlin side.

The initial implementation reads the ac sequence
numbers and allows to mark the client as migrated.

Note: before the migration, glean-ac will need
to land some code to set 'wasMigrated' to false,
always, in case we need to roll back.