Fix the test failure due to DST
Unfortunately, ISO8601 strings that we use as a storage
format in AC do not seem to have a way to encode DST. For
this reason, when loading them off the disk, DST was applied
again. The fix is to set the Java Calendar used to store
the parsed date to the "UTC" timezone.