Take DST into account when converting a calendar into its items
ZONE_OFFSET only provides the default offset from UTC without DST considerations.
DST_OFFSET needs to be added to get the full offset for the calendar instance.

There's a _very small_ hint in the deprecated `Date` class that one
needs to do this[1].

Neither the docs for ZONE_OFFSET[2] nor the ones for DST_OFFSET[3]
mention this explicitly.
I guess the existence of both fields is the only indicator there.

[1]: https://docs.oracle.com/javase/7/docs/api/java/util/Date.html#getTimezoneOffset()
[2]: https://docs.oracle.com/javase/7/docs/api/java/util/Calendar.html#ZONE_OFFSET
[3]: https://docs.oracle.com/javase/7/docs/api/java/util/Calendar.html#DST_OFFSET