Keep track of pings in all modes
We need to keep track of pings, so they get re-registered after a reset.
This state is kept across Glean resets, which should only ever happen in test mode.
Or by the instrumentation tests (`connectedAndroidTest`), which relaunches the application activity,
but not the whole process, meaning globals, such as the ping types, still exist from the old run.

I'm not super happy with keeping this state around, but given that the
queue is actually a set and it's unlikely to have hundreds and hundreds
of custom pings it should not have much of an impact.