1588451, 1530733: Handle dynamic labels in coroutine tasks
At present, the determination of (a) whether there are too many dynamic labels
on a metric, or (b) the label is invalid, are handled synchronously in the
get() operator (in Kotlin, the bracket operator).

This requires the Glean object to already be initialized at metric recording
time -- both to see how many labels have already been used, and also to record
any errors. This both means that labeled metrics can't be recorded prior to
initialization (even though most metrics now can), but more importantly,
blocking I/O is performed to determine how many labels have already been
recorded for the metric.

This PR changes how things work so the requested dynamic label is stored on
the specific labeled metric as a new field. This dynamic label is then converted
to its actual value later as part of the coroutine task, at which point we are
guaranteed to have an initialized Glean object.