Share metrics.yaml and pings.yaml content across platforms.
This will be really handy as we add our third platform (Python).

This also adds a metric for `glean.error.invalid_state`, which wasn't added
when that feature was implemented.  (Missed because that metric is generated
"manually", but we still should have it in the metrics.yaml for documentation
purposes).