Use the zip release of the GitHub release tool
We use ghr, a GitHub release tool, to upload files to GitHub Releases.
For Darwin (macOS) this comes in a zip file (whereas it is a tar.gz for
Linux).
When implementing that initially I copied over the Linux part, but
didn't update the format.
This is changed here now to use the zip file and unzip it.