Enhance `new_glean` test function to attend to all test cases
The `new_glean` function is a helper to create a new glean instance
for tests, to make code less repetitive. It didn't work for cases
when we needed to create multiple glean instances in the same test
with the same data directory. That was fixed by adding the option
to pass the tempdir as an argument to the function, thus making it
possible to call it multiple times with the same data directory.