Add method to persist ping lifetime data on demand
When `delay_ping_lifetime_io` is set, all the create /
read / update / delete methods for Lifetime::Ping data
happen in memory. This commit adds a method to save this
data in bulk on demand to rkv.
  Because of the way that works, I also added  functionality
to load data from disk to memory on creation of a new
instance of Database.