Reset core client metrics when re-enabling upload
We clear all metrics when upload is disabled (glean-core takes care of that),
so we need to reset the initial values only known to the platform when
re-enabling upload.

Now the client info and the channel are passed into the initialize
function of glean_preview.
Before this commit we didn't keep that data around, which means we also
couldn't restore it.
By introducing a small wrapper around the Glean instance we are able to
carry around that data and reuse it when necessary.

This now also adds tests for glean-preview with one big caveat:
Because glean_preview is a global-singleton, we need to run the tests
single-threaded to avoid different tests stomping over each other.
There's no explicit check in the tests, developers will notice problems
by tests just hard crashing on them.

Pass `--test-threads 1` to `cargo test --` or set an environment
variable: `RUST_TEST_THREADS=1`.