Queue ping submission behind other metric operations
Metric operations (e.g. `Counter.add`) that are chronologically run
started before the ping submission is invoked should be included in that
ping (and might actually be the reason the ping finally gets send,
because it would be empty otherwise).
Previously ping submission was always launched on the concurrent queue,
without waiting for the metric operations queue.
This can lead to race conditions in which the ping collection runs
before the metric operations.

Now we put it on the same queue, collect the data (this should be fast
enough) and only put the _sending_ on the concurrent queue.

The `ViewControllerTest` most likely ran into this issue, manifesting
itself by frequent hard-to-reproduce interkittens on CI (indicated by "Asynchronous wait failed: Exceeded timeout of 10 seconds").

The send button of the application records a metric, then sends a ping
that will include the metric.
If that button is pressed right after start of the application no other
data is stored.
If the metric from the button press isn't recorded before the ping is
collected, the ping will be empty and not be sent and the test is
waiting for something that will never occur.