Lint and update sdk_generator.sh script
The now-popular shellcheck tool is a static analysis tool for shell scripts.
It can detect common mistakes or potential bugs in shell scripts.

As the sdk_generator.sh is intended to be used by Glean users, we should
ensure it works across all systems and configurations.
This commit ran shellcheck against it and fixed all warnings and errors:

* Use shell arrays instead of appending to strings
* Properly quote paths to not fail when they contain spaces (not too uncommon on macOS after all)
* Rely on bash for expansion and math