Synchronize glean-preview tests on a global lock
Because glean_preview is a global-singleton, we need to run the tests one-by-one to avoid different tests stomping over each other.
This is only an issue because we're resetting Glean, this cannot happen in normal use of glean-preview.

We use a global lock to force synchronization of all tests, even if run multi-threaded.
This allows us to run without `--test-threads 1`.`