Run CI check for updated schema
Previously we ran schema checks in Python (`test_glean.py`) and Kotlin (several).
These relied on fetching the schema from the mozilla-pipeline-schemas repository.
For reproducability it uses a fixed git hash across both tests.

Now we also bundle a schema into the repository and run a single test
against it from glean-preview.

This is a small imrprovement, but it still means we can miss updating it
(or miss updating all places in the repository to the same version).
The previous commit already introduced a script to automate that.

This CI task now runs an explicit check, failing if the upstream version
changed.

This does mean that unrelated PRs can now fail due to schema changes,
that may not affect them.

In a future world we could have a nightly task that runs this check and
opens a PR with the changed files.