Update locked dependencies
Nothing but running `cargo update`.
We rely on Cargo.lock for reproducible builds as we ship precompiled
libaries for Android and iOS.

We should still be compatible with the versions specified in the
respective Cargo.toml files, which is important for integration into
m-c, where we don't necessarily want to upgrade all transitive
dependencies all the time.