Expire 'user' lifetime metrics
This adds a mechanism to check any 'user' lifetime
metric against an internal list of metric ids and
expiration dates. The expiration date for each metric
is checked against the date the application was started,
not the build date. This is a major different compared to
our build-time expiration mechanism for other lifetimes.
We may live with this, since we don't expect 'user' lifetimes
to have expiration in the long run.