Swift: Add another convenience function to measure fallible functions
Swift is more explicit: fallible functions have to be annotated as such (`throws`).
Everything else can't throw.
If someone wants to measure a fallible function (but get the same
cancellation of the timing when thrown) we need another function.

Luckily we can overload functions, so we can keep the same name.