Bug 1628301, Bug 1628297: Fix race condition deleting tempdir
Glean._reset is called

1) as part of resetting every unit test, both within Glean and for our users
   writing Glean-aware unit tests,

2) atexit

Glean._reset deletes the Glean data directory if it's a temporary directory
(but not under normal circumstances where a permanent location is used).

There is a race condition between the ping uploader and Glean._reset if
the data directory is deleted while pings are still being uploaded. This
solves that by always firing off a single ping uploader process at a time,
and running both ping uploaders and the temporary directory deletion on that
same one-at-a-time process.

It does not use a Process pool to keep a single process around, because
that would add a lot of complexity to handle a daemon process, vs. this
"fire and forget" process.

This also makes Glean._reset private.