Bug 1628310: Don't raise HTTPException or socket.gaierror
These exceptions occur when the network is down, the URL doesn't resolve,
and a bunch of other things. If any of these fail, we should just log and return
False so the ping uploader can move on to other pings or try again next time.
This is preferable to displaying a long (but ultimately harmless) traceback.