[doc only] Bug 1600946: Document how to control logging
The bug proposes integrating the logging level control between Python and Rust,
but I think this has minimal benefit for the amount of complexity.

This just documents that there are two sources of log messages, how to control
them and where to get more info if you need to do something more sophisticated.