Use a bigger integer to store the result
Kotlin is ... not good at handling unsigned numbers. Therefore u16
becomes i16 and thus we quickly run into problems.
We would still have enougb bits to store it all, but we need conversions
on the Kotlin side, because other values are `Int`.
Plus we already mistakenly used `Int` in Kotlin, where it had to be `Short`.

So let's just switch to `Int` (32-bit) and be done with it.