iOS: Build dependencies in Release mode
Otherwise we will get an error about missing bitcode in Gzip:

> Gzip was built without full bitcode.
> All object files and libraries for bitcode must be generated from Xcode Archive or Install build file.

There's no proper workaround short of modifying the dependency's build
system, which we don't want to do.
Bitcode is huge and thus not included in Debug builds by default. It's
required for App Store submissions and thus enabled in Release mode.