flate2 dependency: Don't require oxidized version by default
"miniz_oxide" is the default for flate2, but if Glean is a dependency of
a bigger project that transitively also uses flate2 somewhere else, the
backend in use might change.
We don't really care about the backened in use.

This is fixing an issue when integrating back into mozilla-central,
where flate2 is already in tree and transitively used _somehwere_ with
different features.