Bug 1641564: Add same thread safety guarantees as Kotlin
Python isn't as problematic for threading as Kotlin. For one, the GIL means
there's no collections that completely fail when multithreaded. For another,
multithreading is less common due to the GIL. However, there are still logical
race conditions that could occur in a multithreaded application, and this
applies the same `synchronized` annotations that we have in Kotlin to Python to
prevent those.