Run cbindgen on stable rust (but pretend it's nightly)
`RUSTC_BOOTSTRAP` turns a stable compiler into pretending it is nightly,
unlocking nightly-only features.
It shouldn't be used in a normal build, ever.

But here we _only_ need it for some special nightly-only output flags
for rustc, as used by cbindgen, to determine some metadata about the
crates it looks at.

A recent cargo nightly regression caused that to fail, so this will try
out if we can depend on (pseudo-)stable instead.