Move the C# sample to the main C# solution
This is required in order for the sample application
to correctly reference the local Glean SDK, without
the need to download it from a local nuget. This was
required now since VS often confuses the local and the
remote version of the Glean SDK, if both are available
in the nuget store, even if visual studio is configured
to only fetch from the local store.