Make sure Glean C# doesn't crash on x86
This replaces the use of `SafeHandle`-derived classes
when passing Glean handles back and forth from the FFI
layer. It was crashing on x86 because, while Glean
correctly uses uint64 as handles, `SafeHandle` was
using `IntPtr` under the hood, which is 4 bytes
on 32 bit architectures.

This made glean-core panic because it was seeing an
unknown handle when it was passed back from C# to
glean-core over the FFI.

All in all, we don't really need `SafeHandle`(s) for
everything: we just need it for Win32 handles or
pointers coming from the FFI. Moreover, we don't really
care yet about cleaning up, so no need to get desperate
about re-implementing `SafeHandle`-like mechanism for uint64:
that would also be complicated, because the .NET marshaler
knows about `SafeHandle` and behaves in a special way with
that, when doing P/Invoke on native libraries.