Decouple extraction from validation in glean-core debug opts
Before this PR, validation functions were also performing
the parsing of the data from environment variables. Unfortunately,
by being entangled, it was hard to provide options that were not
of type `String`. This PR fixes the problem by introducing an
`extraction` function: when instantiating a new option, we should
define how we want the data to be extracted from environment
variables.