Allow running non-default activities after the `GleanDebugActivity`
This introduces the `startNext` extra intent parameter for the
`GleanDebugActivity`. Callers can now specify which activity to
run right after this one, instead of the default launch activity.
This is especially useful when debugging telemetry on tests that
are required to run with specific intents or on CI.
See bug 1634064 for more context.