Add the GleanParser task to c# language bindings
This is mostly inspired by this very interesting article:
https://natemcmaster.com/blog/2017/07/05/msbuild-task-in-nuget/

The idea is that this will generate a self-contained Task
that will be referenced by consuming projects at build time.
The task is shipped in the "Glean" main assembly.
Within this repository, we're calling the parser task directly
due to the fact that, relying on assembly files, makes them locked
by msbuild, failing a second build.