Suppress proguard warnings for `mozilla.telemetry.glean.testing.**`
The Glean SDK ships with classes used for tests as well. They are disabled
and not directly usable in production code: they throw if used there. They
can be used in tests just fine but, outside of tests, the test dependency
they use won't be there, hence the warning. It's safe to suppress these.