Work around a crash on Android 8 buggy devices
This works around a bug in the Java compiler on
some Android 8/8.1 devices (SDK=25). See [here](https://issuetracker.google.com/issues/110848122#comment17)
for more details. On buggy devices, an `AssertionError` is
raised because the timezone name `enum` switch is compiled
into racy code. Since we only use the timezone name in
our logs, we can safely swap that out.

We don't need this defensive programming in `DateUtis.kt`
as we're not using timezone names there.