Fix failing android tests
The problem was that the test was trying to get a request sent
by a worker, but that worker received a PingUploadTask::Wait before
getting a PingUploadTask::Upload and the test was trying to check the
request before it had actually arrived.

This is potentially an issue with more tests, but it only manifested
in this test for now. I opened Bug 1657589 for more in depth
investigation on this.