Swift: Ignore Apple Silicon target in Xcode 12
Carthage in default configuration breaks compilation in Xcode 12,
as it tries to build for the simulator now for the arm64 target as well.
That target however is for Apple Silicion, the upcoming Arm platform.
We don't need to run simulator builds there yet (we don't even have
those devices!), so we can safely ignore it.

We need to teach both Carthage about it and for simplification we do the
same for the sample app specifically.

Upstream issue: https://github.com/Carthage/Carthage/issues/3019