Implement a quota based on number of pending pnigs
We limit it to a maximum of 250 pending pings.

The average number of baseline pings per client (on Fenix) is at 15 pings a day.
The P99 value is ~110.
With a maximum of 250 we can store about 2 days worth of pings.
A baseline ping file averages about 600 bytes, so that's a total of just 144 kB we store.
With the default rate limit of 15 pings per 60s it would take roughly 16 minutes to send out all pending
pings.