Document all ping_info fields in the metrics definition file
All of these "metrics" land in the `ping_info` section.
Internally this is considered the `glean_internal_info` ping.

One note: `experiments` is actually a key-value dictionary.
This does not exist as a metric type, we therefore label it a `string`.
The pipeline won't like this when it generates a schema purely based on
what's annoated in metrics.yaml