Don't require the ping name in glean-core testing API traits
This changes the traits in glean-core so that RLB types
can implement testing APIs that do not require developers
to always specify the name of the ping to query in tests.
In line with the other bindings, testing APIs will assume that
the first ping listed in "send_in_pings" should be queried.