Add library-loading configuration hooks for megazording.
The application-services team is experimenting with including Glean
as part of their "megazord" library, a collection of Rust components
that are all compiled together into a single unit. See the PR for
more details:

  https://github.com/mozilla/application-services/pull/3669

This commit adds some library-loading configuration hooks so that Glean
can choose between loading either the standalone `libglean_ffi.so` or
the combined appservices megazord, based on build-time config.