Add test coverage in glean-core for the string type
The test coverage found in the other language bindings,
for example in Kotlin, mostly test that FFI works as
expected. We're not using the FFI in the RLB and most
of the tests should already be covered by the metrics
in glean-core, since RLB only provides a minimal wrapper
to call glean-core within a dispatcher and by acquiring
the Glean singleton.

This specific PR adds test coverage in glean-core for
the string being too long, which was missing.