Bug 1679949: Don't block on running ping uploading process
We only want to run one ping uploading subprocess at a time -- for one, it might
delete ping files, and we can't have contention for that. This is currently
handled by waiting on an existing uploading subprocess before firing off a new
one. However, this can cause the main dispatcher thread to block, causing work
to go uncompleted and subsequent pings to go unsent, on shutdown of the main process.

Instead, this just does nothing if an uploading subprocess is already running,
since the existing uploader will also pick up any new pings added to the
outgoing directory and send them.