Upgrade to newer manylinux image
manylinux10 is based on CentOS 6, which as of 2020-12-02 09:00 is dead.

From http://mirror.centos.org/centos/6/readme:

> The whole CentOS 6 is *dead* and *shouldn't* be used anywhere at *all*

We rely on installing an SSH client (probably to checkout the repository),
so CentOS 6 is no good for us.

manylinux2014 is CentOS 7.8, which is also unmaintained, but at least
continues to work for now.

[doc only]