New crate to generate automatic Rust bindings to the FFI API
We use bindgen to generate the FFI API layer to be used dynamically.
bindgen-generated code then relies on libloading to load a dynamic
library and look up named symbols (`dlsym`).

Currently this will panic if a function is called that cannot be found
(but it only panics on first call, not on initial load).