Kill the dispatcher thread while blocked if Glean is not initialized
There's no point in waiting for it to finish initialization if we're
shutting down. There might be multiple operations queued, which could
block for a long time. We should not hold off shutdown in those cases.

Note: If Glean was already fully initialized shutdown will wait for
these operations to finish before it shuts down.