Use glean-ffi-sys from the RLB and re-export a way to initialize it
The new "dynamic Glean" mode is activated by a compile-time feature.
When enabled consumers need to call `glean::setup_glean` with a path to
the library to load (usually `libglean_ffi`).
Internally the library is loaded and a handle is stored within a global
static.
All metric implementations should then reach for this and convert calls
into `glean_*` FFI calls.