Correctly nest the wait body within the union
As defined previously it was _either_ a tag, an upload body or the wait
time, but of course it needs to be a tag _and_ a wait time (or the
upload body).

So we extract the wait part into its own struct just as we do for the
upload body.
This matches the C definition of it (though the C header now inlines
another struct, because that's valid in C)