Swift: Check and reset dirty flag on startup
This will avoid sending a dirty startup ping if the previous run of the application didn't
successfully and completely initialize Glean.

Unfortunately this is impossible to really test.
Because we currently consider initialization as a "foreground notification" (for the lack of other signals)
we immediately re-set the dirty flag to true.
So once the `initialization` completes we can't observe the previous
state anymore.