Implement timespan dynamically
Glean makes one guarantee:
You can record data at any point and Glean will ensure that data is
stored.
This is true even if you record before Glean is initialized.
However .... in a dynamic world we need to have the dynamic library
loaded before we can construct metrics.
The only way to ensure this in our code is to rely on the dispatcher:
It's only unblocked when the dynamic library is called.
So we now have to defer all metric constructors to initialize through
the dispatcher.

This implements it only for the timespan currently, but needs to be
refactored to be part of the macro, then applied to all metric
implementations.