Delay flushing the dispatcher until after glean-sys is set up
Due to control being on the Kotlin side we need to be careful with the
order in which we load the library and unblock the dispatcher.
Obviously we need to load first, then flush.

But the RLB now automatically handles flushing.
If the Kotlin side calls this early we need to rememeber this, wait for
glean-dynamic to be initialized and then flush as part of the
initialization process.

Howevr if glean-dynamic was initialized _very_ early we can go ahead and
flush right away (however unlikely in Fenix).