bug 1645166 - Support rate external denominators
Rates and Counters are, to the db and to our definitions files, just that:
Rates and Counters.

But to the Language Bindings they can be of a few varieties. Rates can
be RateMetric or NumeratorMetric (for the case of an external denominator).
Counters can be CounterMetric or DenominatorMetric (similarly).

This allows for custom behaviours (CounterMetric can be Labeled,
DenominatorMetric stores its numerators).

Note that Numerator, being a strict subset of Rate, has no Core impl. LBs'
NumeratorMetrics are built atop Core RateMetrics, hiding denominator access.