Use rust compiler host triple instead of sys.platform
The rust host triple is its default target. Instead of relying on the
python platform, we now rely on that host triple to make decisions,
which is more likely to fit what we really want.

And now that we know the default target for the rust compiler, we can
also always explicitly pass it, such that the shared library path is
consistently under target/$target, which simplifies things.

At the same time, we stop assuming "linux" and "darwin" means x86_64.